chore(kms-connector): prepare ethereum listener
