+++ kms-connector/Cargo.toml
@@ -19,7 +19,9 @@ kms-worker.path = "crates/kms-worker"
 tx-sender.path = "crates/tx-sender"
 connector-utils.path = "crates/utils"
 fhevm_gateway_bindings = { path = "../gateway-contracts/rust_bindings", default-features = false }
-fhevm_host_bindings = { path = "../host-contracts/rust_bindings", default-features = false }
+# TODO: use path for host bindings once PR has been merged
+# fhevm_host_bindings = { path = "../host-contracts/rust_bindings", default-features = false }
+fhevm_host_bindings = { git = "https://github.com/zama-ai/fhevm.git", rev = "e2dc53b6a9dcb3fdb8ec92bf1a8157eece7be46c", default-features = false }
 kms-grpc = { git = "https://github.com/zama-ai/kms.git", tag = "v0.13.0-rc.1", default-features = true }
 bc2wrap = { git = "https://github.com/zama-ai/kms.git", tag = "v0.13.0-rc.1", default-features = true }
 tfhe = "=1.5.1"
