+++ kms-connector/crates/gw-listener/Cargo.toml
@@ -10,6 +10,7 @@ version.workspace = true
 bc2wrap.workspace = true
 connector-utils.workspace = true
 fhevm_gateway_bindings.workspace = true
+fhevm_host_bindings.workspace = true
 
 actix-web.workspace = true
 alloy.workspace = true
