+++ kms-connector/crates/gw-listener/src/core/mod.rs
@@ -1,7 +1,10 @@
 mod config;
+mod ethereum;
+mod gateway;
 mod gw_listener;
 mod publish;
 
 pub use config::Config;
-pub use gw_listener::GatewayListener;
+pub use gateway::GatewayListener;
+pub use gw_listener::EventListener;
 pub use publish::publish_event;
